refactor: move surplusModal state to entities
